LPD-47242 Move
