+++ modules/apps/frontend-data-set/frontend-data-set-sample-web/src/main/java/com/liferay/frontend/data/set/sample/web/internal/serializer/FDSSerializerUtil.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.frontend.data.set.sample.web.internal.util;
+package com.liferay.frontend.data.set.sample.web.internal.serializer;
 
 import com.liferay.frontend.data.set.serializer.FDSSerializer;
 import com.liferay.portal.kernel.module.service.Snapshot;
